Extract hero and featured products to client components
Move hero section and featured products section into client components
using useTranslations and useParams instead of server-side
getTranslations + await params. This removes all async operations from
the page component, allowing faster PPR static shell rendering.
